+++ zxcvbn-core/OptimalValues.cs
@@ -16,8 +16,8 @@ namespace Zxcvbn
         {
             for (var i = 0; i < length; i++)
             {
-                G.Add(new Dictionary<int, long>());
-                Pi.Add(new Dictionary<int, long>());
+                G.Add(new Dictionary<int, double>());
+                Pi.Add(new Dictionary<int, double>());
                 M.Add(new Dictionary<int, Match>());
             }
         }
@@ -25,7 +25,7 @@ namespace Zxcvbn
         /// <summary>
         /// Gets or sets the overall metric for the best guess.
         /// </summary>
-        public List<Dictionary<int, long>> G { get; set; } = new List<Dictionary<int, long>>();
+        public List<Dictionary<int, double>> G { get; set; } = new List<Dictionary<int, double>>();
 
         /// <summary>
         /// Gets or sets the best match at a given length.
@@ -35,6 +35,6 @@ namespace Zxcvbn
         /// <summary>
         /// Gets or sets the the product term of the metric for the best guess.
         /// </summary>
-        public List<Dictionary<int, long>> Pi { get; set; } = new List<Dictionary<int, long>>();
+        public List<Dictionary<int, double>> Pi { get; set; } = new List<Dictionary<int, double>>();
     }
 }
