+++ zxcvbn-core/Scoring/BruteForceGuessesCalculator.cs
@@ -11,21 +11,21 @@ namespace Zxcvbn.Scoring
         /// <summary>
         /// The minimum submatch guesses for a multi character string.
         /// </summary>
-        public const int MinSubmatchGuessesMultiCharacter = 50;
+        public const double MinSubmatchGuessesMultiCharacter = 50;
 
         /// <summary>
         /// The minimum submatch guesses for a single character.
         /// </summary>
-        public const int MinSubmatchGuessesSingleCharacter = 10;
+        public const double MinSubmatchGuessesSingleCharacter = 10;
 
-        private const int BruteforceCardinality = 10;
+        private const double BruteforceCardinality = 10;
 
         /// <summary>
         /// Estimates the attempts required to guess the password.
         /// </summary>
         /// <param name="match">The match.</param>
         /// <returns>The guesses estimate.</returns>
-        public static long CalculateGuesses(BruteForceMatch match)
+        public static double CalculateGuesses(BruteForceMatch match)
         {
             var guesses = Math.Pow(BruteforceCardinality, match.Token.Length);
             if (double.IsPositiveInfinity(guesses))
@@ -33,7 +33,7 @@ namespace Zxcvbn.Scoring
 
             var minGuesses = match.Token.Length == 1 ? MinSubmatchGuessesSingleCharacter + 1 : MinSubmatchGuessesMultiCharacter + 1;
 
-            return (long)Math.Max(guesses, minGuesses);
+            return Math.Max(guesses, minGuesses);
         }
     }
 }
